requirements.txt
alembic==1.17.2
argon2-cffi==23.1.0
fastapi==0.136.0
itsdangerous==2.2.0
Jinja2==3.1.6
openpyxl==3.1.5
psycopg2-binary==2.9.11
pydantic==2.13.2
python-dotenv==1.2.2
python-multipart==0.0.26
SQLAlchemy==2.0.49
tzdata==2025.2
uvicorn==0.44.0
